Feature/jis/replace spray with circe everywhere (#1275)
* Rip out Spray JSON everywhere

This commit replaces spray with Circe. The reasons for doing so are that
writing boilerplate is boring, defining jsonFormatX for each case class A
and PaginatedResponse[A] in our datamodel was repetitive and something to
remember all the time, and we weren't happy with the results of spray json
anyway.