Add authorization function
A member of the root organization is a super user. Two methods are
added to the User class: isInRootOrganization, which indicates if
the user is a super user, and isInRootOrSameOrganization, which
gives permissions to both members and super users.

These functions are designed to be used with the AKKA HTTP authorize
directive.