Limit updates to members and super users
This protects from a user sending an HTTP PUT to a resource in
some other organization, but with the user's organizationId in
the request body. Without this, a malicious user could gain
control of someone else's resource by updating its organization
to be that of the users, while passing the authorization check
which only inspects the organizationId of the incoming request
body, not of the resource itself.