Refactor for tile access
Scenes are cached by UUID, and thus need a way to be fetched with
nothing more. Projects need to be checked for tile visibility for
direct tile access.

Two new get methods are added to scenes and projects for supporting
these use cases. The tile access methods are renamed accordingly.