Standardize updates and deletes to single results
We add a CommonHandlers trait, a place to hold handlers that can
be reused in multiple places. completeSingleOrNotFound is the
first such handler, which completes with a 201 if the integer
result is exactly 1, a 404 if it is 0, and a 501 otherwise.

This is then reused in all update and delete instances, since they
all respond identically.

Refs #1365