Improve use of masks when generating footprints
The previous method relied on setting any value in the NIR band less
than 255 as non-missing data and values of 255 as nodata. However,
that was fairly specific to the initial use-case and not
generalizable. This commit switches the method of determining the
datamask using rasterio's builtin `src.dataset_mask` for more robust
nodata detection and mask generation.