Fix reversal of crs and extent_crs
This commit fixes the reversal of extent crs and crs in creating an
ingest definition.

In some ways, this mirrors
https://github.com/azavea/raster-foundry/pull/1184 but on the "client"
side that is generating the ingest definition.

I'm inclined not to write a test case here as we transition away from
using the python library for this sort of thing (for this very reason)
it does not seem very worthwhile.