Add tiff as option for tile export
The tile export is meant for requesting small areas at reasonable zoom
levels. Previously it was possible to request a PNG for these areas,
but it was not possible to request a geotiff with original values for
an area. This commit adds the ability to request PNGs and geotiffs,
with or without color correction.