Increase Airflow parallelism
All of the changes to `docker-compose.airflow.yml` pertain only to the
development environment. They are mainly aimed at keeping proportions closer to
what they are in the default Airflow configuration.

Changes to the Airflow configuration file are to better align with the Airflow
defaults based on the version we are currently running. The only one that should
have a direct impact to AWS environments is `max_threads` for the scheduler.