Explicitly set default_server for Nginx virtual hosts
In scenarios where we know that we want a set of virtual hosts per Nginx
instance, ensure that we set the `default_server` for each port
combination. This will cause Nginx to raise an error in the even that
multiple virtual hosts define that they are the default within a single
instance's configuration.