Adds streamhandler to root logger
This commit resolves a problem where we observed the logs available in
the Airflow UI differed from those we captured by our logging
infrastructure. The problem is that Airflow writes to a file by
default then flushes that file to S3. Ideally we could add a handler
to the Airflow logger and it would also output to stdout. However,
Airflow uses the root logger everywhere, so the only way to capture
the log output is to add a new handler to the root logger which this
commit does.