Check for extent CRS in ingest definition
This set of option checking was a little risky and likely was added
due to some assumptions that held earlier but in retrospect was not a
good set of assumptions. Prior to this commit, we assumed that if an
extent was passed in and the image CRS was defined then the extent CRS
would be lat/lng; however, that isn't necessarily the next few commits
would have introduced a potential case where that was not true.