Split large uploaded Tif Files (#2231)
* Check for extent CRS in ingest definition

This set of option checking was a little risky and likely was added
due to some assumptions that held earlier but in retrospect was not a
good set of assumptions. Prior to this commit, we assumed that if an
extent was passed in and the image CRS was defined then the extent CRS
would be lat/lng; however, that isn't necessarily the next few commits
would have introduced a potential case where that was not true.

* Install GDAL 2.2 via sid repository

This updated GDAL and installs python-gdal for version 2.2 by using
debian's sid repository.

Also, Removes the testing package repository since it was unused and
explicitly uses jessie-backports for the java updates.

* Remove extent, extent_crs, and crs from ingest source definition

This commit removes these attributes from the ingest source definition
because all of this information can be obtained from the tiff tags.
Parsing this information here and potentially later introduced
additional opportunity for errors.

* Add optional tile splitting for large tifs

Adds functions to split tifs larger than 1 gb into component tiles
all associated with a single scene. The split tifs are uploaded back
to the same key prefix of the original file to make the association
clear.

* Use scientific notation for easier number parsing

Makes it easier to understand what the numbers mean

* Use match/case for readability

Switches up the use of the if/else in source definition parsing to be
slightly easier to read and parse for future developers.