--remove-orphans instead
Running the batch jar depends on having the rollbar token, so I threw the .env
file into the airflow worker. However, something in running the airflow test
suite depends on boto, which uses the AWS_PROFILE variable in the .env file if I
throw it in there, and raster-foundry isn't an available profile in the airflow
container in docker-compose.test.