Add ability to override Memcached client mode
When set to static, the set of endpoints specified during initialization
are used throughout the lifetime of the client object. This is what we want in
development because our local Memcached instance does not support
auto-discovery.

When set to dynamic (the DefaultConnectionFactory default) the set of cache node
endpoints, and any updates to it, are dynamically managed. This is what we want
in production because ElastiCache supports auto-discovery.