Handle missing images when scenes' images don't exist in s3 (#2358)
The Landsat 8 importers should understand that not every image file for every
scene is always going to be available (human error might happen, or some images
just might not exist). Instead of failing when that occurs, it should return
some signal value that the image wasn't available and get on with the rest of
what it wants to do. This commit sets that signal value to -1 and lets importer
carry on its merry way.