Use updated assets when downloading planet data
The asset dict for assets from Planet do not include a `location`
attribute until an asset is activated. If an asset happens to be
activated prior to trying to import, everything works ok; however, in
most situations at least one download would fail.

This commit fixes this by returning the updated assets dict with a
location key.

Additionally, an exception is logged with the dict now so that it is
easier to tell if/where something went wrong.