Parse bucket more delicately from upload s3 source (#2427)
* Parse bucket more delicately from upload s3 source
Similar to the problem with prefixes, we were making too strong an assumption
about what we'd be able to parse from the user-input bucket. When a bucket
name contains all underscores rather than dashes, the bucket name is available
under uri.getAuthority rather than uri.getHost. Using getHost caused us to make
a request to s3 without telling it which bucket to look in.