Add ingest handling for Sentinel-2 Jpeg2000 images
Sentinel-2 images couldn't be ingested for at least two reasons. The files were
in jpeg 2000, rather than geotiff, file format, and the bands didn't have
strictly numeric labels. This commit adds processing to convert jpeg 2000s to
tifs for ingest and expands the BandMaps representation to include a label and
an index-on-write, so we can deal with inconvenient labels for bands like "8A".