Retrieve existing Sentinel 2 once before job runs
When importing Sentinel 2 scenes the importer checks that a scene has not
already been added before adding the scene. Before this commit a database call
would be made for each scene and would produce a heavy load on the database. To
get around this problem, this commit instead retrieves a list of scenes first
and uses that for comparisons when doing the import.