Remove DataCells from visible NodeStatistics
Only show stats that are specifically allows in a const array at the top of the
file. This prevents the API from adding extra and messing up the stat node
styling accidentally. Since node size is manually set, we need to explicitly
update the list and the component size at the same time.
Closes #2603