Use a rewrite rule instead of returning 301
Returning a 301 response here breaks CORS preflight requests, which error out on any response besides a 2xx. Additionally, we're unable to blindly return a 2xx response here, because subsequent requests to `/tiles` are redirected, which also violates CORS rules. From https://www.w3.org/TR/cors/#cross-origin-request-with-preflight-0:

"This is the actual request. Apply the make a request steps and observe the request rules below while making the request.

If the response has an HTTP status code of 301, 302, 303, 307, or 308
Apply the cache and network error steps."