Add Shapefile Upload Processing Route for Shapes API
This commit adds a new route (/api/shapes/upload/) that accepts a multipart form
upload with a zipfile in the "name" field for processing. This zipfile is
expected to contain a shapefile with at least one geometry. A shape will created
using the filename for the file uploaded. Users will have the opportunity to
adjust the name of the shape after upload if they desire to do so in the UI.