Add shared/consistent sorting for scene ordering in projects
This commit adds more explicit scene ordering for projects. In cases where
`scene_order` was undefined the sort ordering was non-deterministic. This commit
adds an abstract sort method for `ScenesToProjects` queries that sorts by
`sceneOrder` and then the ID of the scene. In cases where `sceneOrder` is
defined then that will be used for sorting still.