Set ordering of resolvers to ensure JAI dependency resolves correctly
This follows a fix from other projects that have run into this issue where maven
central says it has the dependency (in this case JAI); however, all the links to
the actual files are incorrect. This fixes the problem by explicitly setting the
resolver order to one where the dependency is known to exist.