Add band count validation to local file upload
Now validates that the number of band in each file being uploaded is the
same as the number of bands specified by the datasource.

If the Datasource has no bands, the warning message reflects that.

Malformed GeoTiffs will raise a different warning message, and uploading
will be disabled until the corrupt file is removed.