Set location of export statuses explicitly for exports
Prior to this commit the location of export statuses was created by
combining some hard coded bucket and keys that started to break down
as we were switching to spark local export. This set of
changes (combined) with the previous commit makes the location of the
statuses more explicit by passing in a URI as a parameter to the jobs
rather than requiring the URI be constructed in the same way in 3
different places.