Disable retrying ingests after status is updated to failed
This prevents a scenario where retrying ingests can cause a large
number of jobs to be kicked off because the status was updated to
failed. This change makes it so that in cases where the status is
updated to failed we do not automatically kick off another job and
assume that manual intervention or external retries will handle the
failure.