Catch geometry parsing errors when reading Sentinel 2 scenes
Occasionally Sentinel 2 scenes include empty geometries that fail to
parse. Because all scene inserts occur in a transaction, no scenes end
up added to the database when there is an error.

Also this commit skips adding metadata on the raw data size of the
imagery - this is unused and adds quite a bit of overhead to scene
creation.