Override authQuery for scene queries
Scene queries are special in that the visibility field on the scene
objects themselves determine which are actually public and viewable by
all. This is done as an optimization because we have an inordinate
amount of public scenes that including them in the access control
rules table slows down querying by quite a bit.