Update scene authorization check to use updated logic
In #3547 the logic for querying scenes was updated to improve
performance; however, this meant that the logic to check for scene
viewing permissions used when checking if a user is authorized to view
the scene's datasource or actions needed to be updated as well since
the rules for viewing public scenes was removed from the ACR table.

This commit renames `authQuery` to be more explicit and also updates
other scene routes that check view permissions to use correct logic.

Closes #3561