Revert change to COPY in Dockerfile.api
In a previous PR attempted to force creation of the /srv/dist by
copying over a single file into the directory. However, this causes
errors for actually built versions of the nginx container used in
staging/production. Two fixes for setting up an environment were
included in the previous PR so it seems safe to revert this one.