Add postgres implicit imports back to files
In an attempt to clean up imports in Raster Foundry, some imports for
postgres implicits were accidentally removed in the following commits:
 - 3d183b
 - bb05fd
 - a5e3d8

Apparently, at compile time there is no way for the compiler to
know if the implicits are being used so IntelliJ will remove it
when optimizing the imports. This reverts that specific change.