Update heuristic for determining if clamping/normalizing is needed
Prior to this commit the logic for determining if it was necessary to
clamp was incorrect/not strict enough. If either the minimum or
maximum value in the histogram was between 0 and 255 redistributing
values was not performed. This causes issues when the range of values
includes 0-255, but the maximum greatly exceeds that value.

After this commit both the minimum and maximum of original values must
be between 0 and 255.