Enable template deletion on frontend (#4287)
It adds two angular resources for template action check and for template deletion.
It adds a delete dropdown on each template item. It checks for permissions
with * or DELETE to show the dropdown. It enables template deletion with
the dropdown button click. It fetches the list of templates again afterwards.