Fix shapefile annotations export (#4300)
The annotations shapefile export was being written to an S3 object with
the scheme and bucket name included in the object name. Things worked,
but object names were misleading.

This change makes it so that the scheme and bucket name are no longer
part of the annotations export object name.