Set gdal options for thread safety
According to Grisha: the C API for reading seems to assume that you won't be in
a multi-threaded environment, which makes it so that no matter what we did in
scala, we were still going to hit threading issues in gdal. These options make
it so that we don't hit those.