Ignore errors from integration test failing
If we don't meet the percentile assertion in the gatling test, that step returns
an exit code of 1. That's good! But it also prevents us from sending the results
to s3. That's bad. Since this isn't part of the success or failure of ci in
general, it makes sense that it should report success if it completes the test.