Set gdal options and upgrade gt-server (#4412)
According to Grisha: the C API for reading seems to assume that you won't be in
a multi-threaded environment, which makes it so that no matter what we did in
scala, we were still going to hit threading issues in gdal. Options set in gt-contrib and here
make it so that we don't hit those (as frequently, though we're still seg faulting on initial read sometimes).