Remove Implicit Execution Contexts (#4415)
This removes an implicit context shift and transactor from the
database subproject because it was polluting other parts of the
application when trying to separate out threading. Additionally, this
also provides explicit names for the threads used in the transaction
and connection execution context to make it easier to separate out the
thread pools when looking at the project.