Buffer requests to api/scenes (#4672)
* Add client body buffers to api/scenes request

Users often POST significant amounts of data to api/scenes, which can
cause Nginx to buffer the client body to disk. Increase the size of
the client body buffers for this endpoint to 128k, in line with the 99p
of POST requests.

* Increase api/scenes buffer and report Lambda User-Agent

* Increase client_body_buffer_size to 256k in the Nginx conf for
  /api/scenes to match 99p request body sizes
* Add a User-Agent header to the Lambda scene creation call in order
  to better track where those requests are coming from

* Report version number of Lambda scene creation function

When the Lambda function creates scenes for Sentinel and Landsat
imagery, pass the version of the function to the request via
the User-Agent header.

* Update CHANGELOG.md

Update the changelog to reference changes made to Nginx buffer sizes
for the scene creation endpoint.

* Report git SHA in Sentinel/Landsat Lambda execution

When POSTing new scenes to the API from a Lambda function, report
the SHA of the commit in the User-Agent header to track the version
of the function being executed.

* Install updated version of six

At least 1.11 is required by `jsonschema`. Building of the `batch`
container image fail otherwise.