Make tile size for exports configurable + fix nodata handling
Two changes are in this particular commit. First, instead of using a
zoomed layout scheme, this commit uses a floating layout scheme that
allows us to more easily specify the tile size of the export. This is
in an attempt to reduce the number of requests + tiles written
out. The second change here is around nodata handling to fix merging
of scenes for exports that were not previously handled properly.