Reorganize subprojects (split out datamodel from common)
This commit does some subproject reorganization to make the dependency
hierarchy and jar sizes more sane. It is now more realistic to depend
_only_ on the datamodel subproject and not necessarily pull in a bunch
of extraneous dependencies.

Additionally, it adds functionality for generating project layer overiews
with the intention of either (a) running it from the command line
or (b) running it in a lambda function.

The default assembly option is to exclude the scala language library
so if running it locally an option must be set in the `sbt` shell to
switch this setting before generating the assembly jar:

`set assemblyOption in assembly := (assemblyOption in
assembly).value.copy(includeScala = true)`