Remove recover from sick and use attempts to get better errors
This switches the cache as well, mostly because we don't use the tile
cache and I wanted to use one that we use more often at this point --
though both point to the same memcached instance so it shouldn't
matter.

The switch to using `attempt` is an _attempt_ at getting better/more
errors and logging them correctly when an error occurs.