Use Separate Pools for Raster IO + Cache
In #5200 we removed `parTraverse` because we had evidence that tiles
with many IO operations were using up the default execution context
threads and blocking on requests from S3. Before, we experimented with
different execution contexts for the whole application, but not in
specific circumstances. This commit introduces `parTraverse` back into
our rendering and specifically uses a cached thread pool in this
context.

Additionally, the memcached client had an internal pool that was
scaled based on the number of cores present. This PR adjusts that to
also use a cached threadpool with the understanding that since this is
IO we should not be limiting ourselves to the cores of the machine.