Set S3 URI to contain original filename
The approach here is to keep the filename from the original upload in
the S3 URI where the file is written so that it can be "carried through"
later on during image processing (if the file already is a valid COG
then this file name will be maintained as per the current image
processing logic).

This change assumes that the signed URL generated for this upload only
contains a single image (this seems to already be an assumption since
the generated URL is only unique per upload id, with the assumption that
each file has its own upload id; if files shared an upload id they would
overwrite each other). Given that, we grab the first available filename
and use that.